Refactor ethereum controllers
Move status report functionality into a separate controller. Move the previous EthereumUIController code into EthereumController. Rename the "ajaxOperation" method to "performOperation" - this was done because the method itself actually functions to perform an operation, and the ajax functionality is just an optional part of it.
